+++ burnxp/src/lib.rs
@@ -3,6 +3,6 @@ mod model;
 mod predict;
 mod train;
 
-pub use model::{ResNetType, ScoreModelConfig};
+pub use model::{ModelConfig, ResNetType};
 pub use predict::{predict, Output, PredictConfig};
 pub use train::{train, TrainingConfig};