+++ burnxp/src/model.rs
@@ -4,18 +4,19 @@ use burn::tensor::backend::AutodiffBackend;
 use burn::train::{MultiLabelClassificationOutput, TrainOutput, TrainStep, ValidStep};
 use clap::builder::OsStr;
 use clap::ValueEnum;
-use nn::loss::BinaryCrossEntropyLossConfig;
+use nn::loss::{BinaryCrossEntropyLoss, BinaryCrossEntropyLossConfig};
 use nn::Sigmoid;
 use resnet_burn::ResNet;
 use serde::{Deserialize, Serialize};
 
 #[derive(Module, Debug)]
-pub(crate) struct ScoreModel<B: Backend> {
+pub(crate) struct Model<B: Backend> {
     resnet: ResNet<B>,
     sigmoid: Sigmoid,
+    loss: BinaryCrossEntropyLoss<B>,
 }
 
-impl<B: Backend> ScoreModel<B> {
+impl<B: Backend> Model<B> {
     /// # Shapes
     ///   - Images [batch_size, 1024, 1024, 3]
     ///   - Output [batch_size, num_classes]
@@ -29,36 +30,32 @@ impl<B: Backend> ScoreModel<B> {
         batch: ImageBatch<B>,
     ) -> MultiLabelClassificationOutput<B> {
         let output = self.forward(batch.datas);
-        let loss = BinaryCrossEntropyLossConfig::new()
-            .with_smoothing(Some(0.1))
-            .init(&B::Device::default())
-            .forward(output.clone(), batch.targets.clone());
+        let loss = self.loss.forward(output.clone(), batch.targets.clone());
         MultiLabelClassificationOutput::new(loss, output, batch.targets)
     }
 }
 
-impl<B: AutodiffBackend> TrainStep<ImageBatch<B>, MultiLabelClassificationOutput<B>>
-    for ScoreModel<B>
-{
+impl<B: AutodiffBackend> TrainStep<ImageBatch<B>, MultiLabelClassificationOutput<B>> for Model<B> {
     fn step(&self, batch: ImageBatch<B>) -> TrainOutput<MultiLabelClassificationOutput<B>> {
         let reg = self.forward_multilabelclassification(batch);
         TrainOutput::new(self, reg.loss.backward(), reg)
     }
 }
 
-impl<B: Backend> ValidStep<ImageBatch<B>, MultiLabelClassificationOutput<B>> for ScoreModel<B> {
+impl<B: Backend> ValidStep<ImageBatch<B>, MultiLabelClassificationOutput<B>> for Model<B> {
     fn step(&self, batch: ImageBatch<B>) -> MultiLabelClassificationOutput<B> {
         self.forward_multilabelclassification(batch)
     }
 }
 
 #[derive(Config, Debug)]
-pub struct ScoreModelConfig {
+pub struct ModelConfig {
     rnn_type: ResNetType,
+    weights: Option<Vec<f32>>,
 }
 
-impl ScoreModelConfig {
-    pub(crate) fn init<B: Backend>(&self, device: &B::Device, num_classes: usize) -> ScoreModel<B> {
+impl ModelConfig {
+    pub(crate) fn init<B: Backend>(&self, device: &B::Device, num_classes: usize) -> Model<B> {
         let resnet = match self.rnn_type {
             ResNetType::Layer18 => ResNet::resnet18(num_classes, device),
             ResNetType::Layer34 => ResNet::resnet34(num_classes, device),
@@ -66,9 +63,13 @@ impl ScoreModelConfig {
             ResNetType::Layer101 => ResNet::resnet101(num_classes, device),
             ResNetType::Layer152 => ResNet::resnet152(num_classes, device),
         };
-        ScoreModel {
+        Model {
             resnet,
             sigmoid: Sigmoid,
+            loss: BinaryCrossEntropyLossConfig::new()
+                .with_smoothing(Some(0.1))
+                .with_weights(self.weights.clone())
+                .init(device),
         }
     }
 }