+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 - improve: self-define hamming score confidence threshold when training
 - improve: more soft lr scheduler
 - improve: more reasonable early stop strategy
+- improve: BinaryCrossEntropyLoss with label smoothing
 
 ## [0.1.0-alpha42] - 2024-10-07