+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 -->
 
 ## [Unreleased]
+## [0.1.0-alpha44] - 2024-10-08
+
+- improve: take label weight into account when training
+
 ## [0.1.0-alpha43] - 2024-10-08
 
 - make observer optional